+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Printing function calls with a single table or string parameter
 - Local variables without immediate assignment crashed the parser
 - The initial minification used reserved keywords (in particular: or)
+- Handling carts with multiple CODE chunks
 
 ## [1.2.0] - 2023-04-02